Add -r/--total-results to limit search results
-r/--total-results allows the user to specify the number of links to return in the search. Because this option operates on the total search results, i.e. it's not the number of links to show per page, -p/--pages is ignored if --total-results is set. Perhaps in the future this option can be extended to operate by page.

Added total-results to config

total-results is set in the config to be 50 by default.
I also reverted the logic that only one page would be fetched if total-results is set. Now it has the original behavior: any number of pages can be requested, and total-results will filter the final result.

Minor wording update

Wording change also for cli parameter